Create 07_after_1_substition_dot_product_multiple.py
You are given an array a of length n. The characteristic of this array is the value  — the sum of the products of the values ai by i. One may perform the following operation exactly once: pick some element of the array and move to any position. In particular, it's allowed to move the element to the beginning or to the end of the array. Also, it's allowed to put it back to the initial position. The goal is to get the array with the maximum possible value of characteristic.


Input
The first line of the input contains a single integer n (2 ≤ n ≤ 200 000) — the size of the array a.

The second line contains n integers ai (1 ≤ i ≤ n, |ai| ≤ 1 000 000) — the elements of the array a.

Output
Print a single integer — the maximum possible value of characteristic of a that can be obtained by performing no more than one move.

Examples
inputCopy
4
4 3 2 5
outputCopy
39
inputCopy
5
1 1 2 7 1
outputCopy
49


explanation of above input and output
In the first sample, one may pick the first element and place it before the third (before 5). Thus, the answer will be 3·1 + 2·2 + 4·3 + 5·4 = 39.

In the second sample, one may pick the fifth element of the array and place it before the third. The answer will be 1·1 + 1·2 + 1·3 + 2·4 + 7·5 = 49.