Update 24_levenshtein_distance.py
Find the Levenshtein Distance between two strings to transfer one string to another
For example, from "test" to "test" the Levenshtein distance is 0 because both the source and
target strings are identical. No transformations are needed. In contrast, from "test" to "team"
the Levenshtein distance is 2 - two substitutions have to be done to turn "test" in to "team".